Fixed 32-64bit multiplcation bug in get_microseconds, and changed some assertions in pause_for_duration.
git-svn-id: http://svn.bramp.net:8080/svn/PhD/Work/userspace/threadnetperf@2732 e02701c3-087c-644a-a2a3-551e05f81084

+++ common.c
@@ -152,7 +152,7 @@ unsigned long long get_microseconds() {
 		fprintf(stderr, "%s:%d gettimeofday() error (%d) %s\n", __FILE__, __LINE__, ERRNO, strerror(ERRNO));
 		microseconds = 0;
 	} else {
-		microseconds = tv.tv_sec * 1000000 + tv.tv_usec;
+		microseconds = (unsigned long long)tv.tv_sec * 1000000 + (unsigned long long)tv.tv_usec;
 	}
 
 #endif
